feat: move terms of service operations to users-endpoint
