+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Activities/Productizer/ProductizerController.cs
@@ -2,12 +2,11 @@ using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
-using VirtualFinland.UserAPI.Activities.Identity.Operations;
 using VirtualFinland.UserAPI.Activities.Productizer.Operations;
 using VirtualFinland.UserAPI.Activities.Productizer.Operations.BasicInformation;
 using VirtualFinland.UserAPI.Activities.Productizer.Operations.JobApplicantProfile;
-using VirtualFinland.UserAPI.Activities.Productizer.Operations.TermsOfServiceAgreement;
 using VirtualFinland.UserAPI.Exceptions;
+using VirtualFinland.UserAPI.Helpers;
 using VirtualFinland.UserAPI.Helpers.Services;
 
 namespace VirtualFinland.UserAPI.Activities.Productizer;
@@ -17,13 +16,11 @@ namespace VirtualFinland.UserAPI.Activities.Productizer;
 [Authorize(Policy = "RequestFromDataspace")]
 [ProducesResponseType(StatusCodes.Status401Unauthorized)]
 [Produces("application/json")]
-public class ProductizerController : ControllerBase
+public class ProductizerController : ApiControllerBase
 {
-    private readonly AuthenticationService _authenticationService;
     private readonly TestbedConsentSecurityService _consentSecurityService;
     private readonly ILogger<ProductizerController> _logger;
 
-    private readonly IMediator _mediator;
     private readonly string _userProfileDataSourceURI;
 
     public ProductizerController(
@@ -31,10 +28,8 @@ public class ProductizerController : ControllerBase
         AuthenticationService authenticationService,
         TestbedConsentSecurityService consentSecurityService,
         ILogger<ProductizerController> logger,
-        IConfiguration configuration)
+        IConfiguration configuration) : base(mediator, authenticationService)
     {
-        _mediator = mediator;
-        _authenticationService = authenticationService;
         _consentSecurityService = consentSecurityService;
         _logger = logger;
         _userProfileDataSourceURI = configuration["ConsentDataSources:UserProfile"] ?? throw new ArgumentNullException("ConsentDataSources:UserProfile");
@@ -117,69 +112,5 @@ public class ProductizerController : ControllerBase
     {
         command.SetAuth(await GetPersonIdOrCreateNewPersonWithId());
         return Ok(await _mediator.Send(command));
-    }
-
-    [HttpPost("productizer/Service/Terms/Agreement_v0.1")]
-    [HttpPost("productizer/Service/Terms/Agreement_v1.0")]
-    [SwaggerOperation(Summary = "Get the user terms agreement status (Testbed Productizer)",
-        Description = "Returns the current logged user terms agreement status.")]
-    [ProducesResponseType(typeof(GetUser.User), StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status404NotFound)]
-    [ProducesErrorResponseType(typeof(ProblemDetails))]
-    public async Task<IActionResult> GetPersonTermsOfServiceAgreement()
-    {
-        var personId = await GetPersonIdOrCreateNewPersonWithId();
-        return Ok(await _mediator.Send(new GetPersonServiceTermsAgreement.Query(personId)));
-    }
-
-    [HttpPost("productizer/Service/Terms/Agreement/Write_v0.1")]
-    [HttpPost("productizer/Service/Terms/Agreement/Write_v1.0")]
-    [SwaggerOperation(Summary = "Update the current logged user terms agreement status (Testbed Productizer)",
-        Description = "Updates the current logged user terms agreement status.")]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesErrorResponseType(typeof(ProblemDetails))]
-    public async Task<IActionResult> UpdatePersonTermsOfServiceAgreement(UpdatePersonServiceTermsAgreement.Command command)
-    {
-        var personId = await _authenticationService.GetCurrentUserId(Request, false);
-        command.SetAuth(personId);
-        return Ok(await _mediator.Send(command));
-    }
-
-    /// <summary>
-    ///     If user is not found in database, create new user and return users Id
-    ///     - authentication header / token should be verified before calling this method
-    /// </summary>
-    private async Task<Guid> GetPersonIdOrCreateNewPersonWithId()
-    {
-        Guid personId;
-        try
-        {
-            personId = await _authenticationService.GetCurrentUserId(Request);
-        }
-        catch (Exception e)
-        {
-            if (e is not NotFoundException && e is not NotAuthorizedException)
-            {
-                throw;
-            }
-
-            _logger.LogInformation("Could not get personId for user with error message: {Error}. Try create new user",
-                e.Message);
-            try
-            {
-                var jwkToken = await _authenticationService.ParseAuthenticationHeader(Request);
-                var query = new VerifyIdentityUser.Query(jwkToken.UserId, jwkToken.Issuer);
-                var createdUser = await _mediator.Send(query);
-                personId = createdUser.Id;
-                _logger.LogInformation("New user was created with Id: {UserId}", personId);
-            }
-            catch (Exception exception)
-            {
-                _logger.LogError("Could not create new user. Error message: {Error}", exception.Message);
-                throw;
-            }
-        }
-
-        return personId;
     }
 }
