+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/appsettings.local.json
@@ -30,6 +30,9 @@
       "SuomiFi": {
         "IsEnabled": true
       }
+    },
+    "Options": {
+      "TermsOfServiceAgreementRequired": true
     }
   }
 }
